feat: add token modal common form and spacing (#1532)
* feat: common form and spacing

* feat: extract toolbar
